refactor: group riven components
