+++ src/App.tsx
@@ -71,8 +71,8 @@ import Foundry from "./Foundry";
 import CacheStatusChip from "./CacheStatusChip";
 import MarketHelper from "./MarketHelper";
 import RelicHelper from "./RelicHelper";
-import RivenAnalyzer from "./RivenAnalyzer";
-import RivenOverlayWindow from "./RivenOverlayWindow";
+import RivenAnalyzer from "./riven/RivenAnalyzer";
+import RivenOverlayWindow from "./riven/RivenOverlayWindow";
 import RelicPickOverlay from "./RelicPickOverlay";
 import TimerHelper, { fmtMs } from "./TimerHelper";
 import { useWorldState } from "./worldstate";
