+++ src/riven/RivenAnalyzer.tsx
@@ -1,10 +1,10 @@
 import { useState, useEffect, useCallback, useRef } from "react";
 import { invoke } from "@tauri-apps/api/core";
 import { listen } from "@tauri-apps/api/event";
-import { checkRivenNow } from "./App";
-import { TAURI_COMMANDS, TAURI_EVENTS } from "./constants/tauri";
-import type { RivenAnalysis, RivenStat, SavedRiven } from "./types/rivens";
-import type { AnalyzeRivenArgs, SaveRivenRollArgs } from "./types/tauri";
+import { checkRivenNow } from "../App";
+import { TAURI_COMMANDS, TAURI_EVENTS } from "../constants/tauri";
+import type { RivenAnalysis, RivenStat, SavedRiven } from "../types/rivens";
+import type { AnalyzeRivenArgs, SaveRivenRollArgs } from "../types/tauri";
 import "./RivenAnalyzer.css";
 
 // ── Types ─────────────────────────────────────────────────────────────────────
